--- src/main/java/com/mumulcom/mumulcom/src/question/dao/MyQuestionListDao.java
@@ -1,54 +1,0 @@
-package com.mumulcom.mumulcom.src.question.dao;
-
-import com.mumulcom.mumulcom.src.question.domain.MyQuestionListRes;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.stereotype.Repository;
-
-import javax.sql.DataSource;
-import java.util.List;
-
-@Repository
-public class MyQuestionListDao {
-    private JdbcTemplate jdbcTemplate;
-
-    @Autowired
-    public void setDataSource(DataSource dataSource) {
-        this.jdbcTemplate = new JdbcTemplate(dataSource);
-    }
-
-    public List<MyQuestionListRes> myQuestionListRes (int userIdx) {
-        String myQuestionListQuery = "select * \n" +
-                "from\n" +
-                "(select u.userIdx, q.questionIdx, nickname, bigCategoryName, smallCategoryName, title, q.createdAt as updatedAt, likeCount, replyCount\n" +
-                "from Question q, BigCategory b, SmallCategory s , User u, CodeQuestion c ,\n" +
-                "(select q.questionIdx, ifnull(likeCount,0) as likeCount\n" +
-                "from Question q left join (select questionIdx, count(*) likeCount from `Like` group by questionIdx) l on  q.questionIdx = l.questionIdx) l,\n" +
-                "(select q.questionIdx, ifnull(replyCount,0) as replyCount\n" +
-                "from Question q left join (select questionIdx, count(*) replyCount from Reply group by questionIdx) r on q.questionIdx = r.questionIdx) r\n" +
-                "where q.bigCategoryIdx = b.bigCategoryIdx and q.smallCategoryIdx = s.smallCategoryIdx and u.userIdx = q.userIdx and c.questionIdx = q.questionIdx \n" +
-                "and q.questionIdx = l.questionIdx and q.questionIdx = r.questionIdx\n" +
-                "union\n" +
-                "select u.userIdx, q.questionIdx,nickname, bigCategoryName, smallCategoryName, title, q.createdAt as updatedAt, likeCount, replyCount\n" +
-                "from Question q, BigCategory b, SmallCategory s , User u, ConceptQuestion c ,\n" +
-                "(select q.questionIdx, ifnull(likeCount,0) as likeCount\n" +
-                "from Question q left join (select questionIdx, count(*) likeCount from `Like` group by questionIdx) l on  q.questionIdx = l.questionIdx) l,\n" +
-                "(select q.questionIdx, ifnull(replyCount,0) as replyCount\n" +
-                "from Question q left join (select questionIdx, count(*) replyCount from Reply group by questionIdx) r on q.questionIdx = r.questionIdx) r\n" +
-                "where q.bigCategoryIdx = b.bigCategoryIdx and q.smallCategoryIdx = s.smallCategoryIdx and u.userIdx = q.userIdx and c.questionIdx = q.questionIdx \n" +
-                "and q.questionIdx = l.questionIdx and q.questionIdx = r.questionIdx) questionList\n" +
-                "where userIdx = ?";
-        return this.jdbcTemplate.query(myQuestionListQuery,
-                (rs,rowNum) -> new MyQuestionListRes(
-                        rs.getLong("userIdx"),
-                        rs.getLong("questionIdx"),
-                        rs.getString("nickname"),
-                        rs.getString("bigCategoryName"),
-                        rs.getString("smallCategoryName"),
-                        rs.getString("title"),
-                        rs.getString("updatedAt"),
-                        rs.getInt("likeCount"),
-                        rs.getInt("replyCount")
-                ),userIdx);
-    }
-}
