+++ src/main/java/com/mumulcom/mumulcom/src/question/provider/QuestionProvider.java
@@ -6,6 +6,9 @@ import com.mumulcom.mumulcom.config.BaseException;
 import com.mumulcom.mumulcom.config.BaseResponseStatus;
 import com.mumulcom.mumulcom.config.BaseResponseStatus.*;
 import com.mumulcom.mumulcom.src.question.dao.QuestionDao;
+import com.mumulcom.mumulcom.src.question.domain.MyQuestionListRes;
+import com.mumulcom.mumulcom.src.question.domain.SearchCodingQuestionRes;
+import com.mumulcom.mumulcom.src.question.domain.SearchConceptQuestionRes;
 import com.mumulcom.mumulcom.src.question.dto.GetRecQueRes;
 import com.mumulcom.mumulcom.utils.JwtService;
         import org.slf4j.Logger;
@@ -14,6 +17,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+
+import static com.mumulcom.mumulcom.config.BaseResponseStatus.DATABASE_ERROR;
 //import com.example.demo.utils.*;
 //import com.example.demo.config.secret.*;
 
@@ -66,5 +71,46 @@ public class QuestionProvider {
       }
    }
 
+   /**
+    * 휘정
+    * 개념 질문 검색하기
+    * */
+   public List<SearchConceptQuestionRes> searchConceptQuestionResList(String keyword) throws BaseException {
+      try {
+         List<SearchConceptQuestionRes> searchConceptQuestionResList = questionDao.searchConceptQuestionRes(keyword);
+         return searchConceptQuestionResList;
+      } catch (Exception exception) {
+         exception.printStackTrace();
+         throw new BaseException(DATABASE_ERROR);
+      }
+   }
+
+   /**
+    * 휘정
+    * 코딩 질문 검색하기
+    * */
+   public List<SearchCodingQuestionRes> searchCodingQuestionResList(String keyword) throws BaseException {
+      try {
+         List<SearchCodingQuestionRes> searchCodingQuestionResList = questionDao.searchCodingQuestionRes(keyword);
+         return searchCodingQuestionResList;
+      } catch (Exception exception) {
+         exception.printStackTrace();
+         throw new BaseException(DATABASE_ERROR);
+      }
+   }
+
+   /**
+    * 휘정
+    * 내가 한 질문 목록 받아오는 API
+    * */
+   public List<MyQuestionListRes> myQuestionListResList (int userIdx) throws BaseException {
+      try {
+         List<MyQuestionListRes> myQuestionListResList = questionDao.myQuestionListRes(userIdx);
+         return myQuestionListResList;
+      } catch (Exception exception) {
+         exception.printStackTrace();
+         throw new BaseException(DATABASE_ERROR);
+      }
+   }
 
 }
